+++ src/main/java/com/gitsh01/libertyvillagers/cmds/VillagerStats.java
@@ -197,7 +197,7 @@ public class VillagerStats {
         return pageString;
     }
 
-    protected static String translatedProfession(VillagerProfession profession) {
+    public static String translatedProfession(VillagerProfession profession) {
         String villagerTranslationKey = EntityType.VILLAGER.getTranslationKey();
         return
                 new TranslatableText(villagerTranslationKey + "." + Registry.VILLAGER_PROFESSION.getId(profession).getPath()).getString();