+++ src/main/java/com/gitsh01/libertyvillagers/config/VillagersGeneralConfig.java
@@ -28,6 +28,12 @@ public class VillagersGeneralConfig implements ConfigData {
     @ConfigEntry.Gui.Tooltip(count = 2)
     public boolean villagerWanderingFix = true;
 
+    @ConfigEntry.Gui.Tooltip(count = 2)
+    public int walkTowardsTaskMaxRunTime = 2400;
+
+    @ConfigEntry.Gui.Tooltip(count = 2)
+    public int walkTowardsTaskMinCompletionRange = 3;
+
     @ConfigEntry.Gui.Tooltip
     public boolean villagersDontClimb = true;
 
@@ -45,6 +51,12 @@ public class VillagersGeneralConfig implements ConfigData {
 
     @ConfigEntry.Gui.Tooltip
     public boolean villagersEatMelons = true;
+
+    @ConfigEntry.Gui.Tooltip
+    public boolean villagersEatPumpkinPie = true;
+
+    @ConfigEntry.Gui.Tooltip
+    public boolean villagersDontTrampleCrops = true;
 
     @ConfigEntry.Gui.Tooltip
     public boolean noNitwitVillagers = false;