+++ src/main/java/com/gitsh01/libertyvillagers/mixin/FarmerVillagerTaskMixin.java
@@ -1,6 +1,7 @@
 package com.gitsh01.libertyvillagers.mixin;
 
 import com.google.common.collect.ImmutableMap;
+import com.google.common.collect.ImmutableSet;
 import net.minecraft.block.*;
 import net.minecraft.entity.ai.brain.BlockPosLookTarget;
 import net.minecraft.entity.ai.brain.MemoryModuleType;
@@ -35,6 +36,8 @@ public abstract class FarmerVillagerTaskMixin extends Task<VillagerEntity> {
     public FarmerVillagerTaskMixin() {
         super(ImmutableMap.of());
     }
+
+    private static final int MAX_RUN_TIME = 20 * 60; // One minute.
 
     @Shadow
     @Nullable
@@ -80,7 +83,6 @@ public abstract class FarmerVillagerTaskMixin extends Task<VillagerEntity> {
                     }
                 }
             }
-
             this.currentTarget = this.chooseRandomTarget(serverWorld);
             cir.setReturnValue(this.currentTarget != null);
             cir.cancel();
@@ -89,19 +91,34 @@ public abstract class FarmerVillagerTaskMixin extends Task<VillagerEntity> {
 
     @Inject(method = "keepRunning", at = @At(value = "HEAD"), cancellable = true)
     protected void keepRunning(ServerWorld serverWorld, VillagerEntity villagerEntity, long l, CallbackInfo cir) {
-        if (!CONFIG.villagersProfessionConfig.preferPlantSameCrop) {
+        if (!CONFIG.villagersProfessionConfig.preferPlantSameCrop &&
+            !CONFIG.villagersProfessionConfig.farmersHarvestPumpkins &&
+            !CONFIG.villagersProfessionConfig.farmersHarvestMelons) {
             // Use default logic.
+            return;
+        }
+
+        if ( villagerEntity.getBrain().hasMemoryModule(MemoryModuleType.WALK_TARGET)) {
+            // Wait for the villager to reach the walk target.
+            cir.cancel();
             return;
         }
 
         Item preferredSeeds = null;
         BlockPos currentTarget = this.currentTarget;
-        if (currentTarget == null || currentTarget.isWithinDistance(villagerEntity.getPos(), 1.0)) {
-            if (currentTarget != null && l > this.nextResponseTime) {
-                BlockState blockState = serverWorld.getBlockState(currentTarget);
-                Block block = blockState.getBlock();
-                Block block2 = serverWorld.getBlockState(currentTarget.down()).getBlock();
+        int distance = 1;
+        BlockState blockState = serverWorld.getBlockState(currentTarget);
+        Block block = blockState.getBlock();
+        Block block2 = serverWorld.getBlockState(currentTarget.down()).getBlock();
+        if (block instanceof GourdBlock) {
+            // Can't stand directly on top of the gourd bottom, so increase the distance.
+            distance = 2;
+        }
+        if (currentTarget.isWithinDistance(villagerEntity.getPos(), distance)) {
+            boolean foundBlockCrop = false;
+            if (CONFIG.villagersProfessionConfig.preferPlantSameCrop) {
                 if (block instanceof CropBlock && ((CropBlock) block).isMature(blockState)) {
+                    foundBlockCrop = true;
                     if (block instanceof BeetrootsBlock) {
                         preferredSeeds = Items.BEETROOT_SEEDS;
                     } else if (block instanceof PotatoesBlock) {
@@ -111,57 +128,86 @@ public abstract class FarmerVillagerTaskMixin extends Task<VillagerEntity> {
                     } else {
                         preferredSeeds = Items.WHEAT_SEEDS;
                     }
-
-                    serverWorld.breakBlock(currentTarget, true, villagerEntity);
-                    blockState = serverWorld.getBlockState(currentTarget);
-                    block = blockState.getBlock();
                 }
+            }
 
-                if (blockState.isAir() && block2 instanceof FarmlandBlock && villagerEntity.hasSeedToPlant()) {
-                    SimpleInventory simpleInventory = villagerEntity.getInventory();
+            if (block instanceof GourdBlock) {
+                if (CONFIG.villagersProfessionConfig.farmersHarvestMelons && block instanceof MelonBlock) {
+                    foundBlockCrop = true;
+                }
+                if (CONFIG.villagersProfessionConfig.farmersHarvestPumpkins && block instanceof PumpkinBlock) {
+                    foundBlockCrop = true;
+                }
+            }
 
-                    // First look for the preferred seed.
-                    boolean plantedPreferredSeeds = false;
-                    if (preferredSeeds != null) {
-                        for (int i = 0; i < simpleInventory.size(); ++i) {
-                            ItemStack itemStack = simpleInventory.getStack(i);
-                            if (!itemStack.isEmpty() && itemStack.isOf(preferredSeeds)) {
-                                plantedPreferredSeeds = plantSeed(itemStack, i, serverWorld, villagerEntity);
-                                if (plantedPreferredSeeds) {
-                                    break;
-                                }
+            if (foundBlockCrop) {
+                serverWorld.breakBlock(currentTarget, true, villagerEntity);
+                blockState = serverWorld.getBlockState(currentTarget);
+            }
+
+            if (blockState.isAir() && block2 instanceof FarmlandBlock && villagerEntity.hasSeedToPlant()) {
+                SimpleInventory simpleInventory = villagerEntity.getInventory();
+
+                // First look for the preferred seed.
+                boolean plantedPreferredSeeds = false;
+                if (preferredSeeds != null) {
+                    for (int i = 0; i < simpleInventory.size(); ++i) {
+                        ItemStack itemStack = simpleInventory.getStack(i);
+                        if (!itemStack.isEmpty() && itemStack.isOf(preferredSeeds)) {
+                            plantedPreferredSeeds = plantSeed(itemStack, i, serverWorld, villagerEntity);
+                            if (plantedPreferredSeeds) {
+                                break;
                             }
                         }
                     }
-
-                    if (!plantedPreferredSeeds) {
-                        // Look for any seed to plant.
-                        for (int i = 0; i < simpleInventory.size(); ++i) {
-                            ItemStack itemStack = simpleInventory.getStack(i);
-                            if (!itemStack.isEmpty()) {
-                                if (plantSeed(itemStack, i, serverWorld, villagerEntity)) {
-                                    break;
-                                }
-                            }
-                        }
-                    }
                 }
 
-                if (block instanceof CropBlock && !((CropBlock) block).isMature(blockState)) {
-                    this.targetPositions.remove(currentTarget);
-                    this.currentTarget = this.chooseRandomTarget(serverWorld);
-                    if (this.currentTarget != null) {
-                        this.nextResponseTime = l + 20L;
-                        villagerEntity.getBrain().remember(MemoryModuleType.WALK_TARGET,
-                                new WalkTarget(new BlockPosLookTarget(this.currentTarget), 0.5F, 1));
-                        villagerEntity.getBrain()
-                                .remember(MemoryModuleType.LOOK_TARGET, new BlockPosLookTarget(this.currentTarget));
+                if (!plantedPreferredSeeds) {
+                    // Look for any seed to plant.
+                    for (int i = 0; i < simpleInventory.size(); ++i) {
+                        ItemStack itemStack = simpleInventory.getStack(i);
+                        // Plant pumpkin seeds over pumpkins if possible.
+                        if (CONFIG.villagersProfessionConfig.farmersHarvestPumpkins && itemStack.isOf(Items.PUMPKIN)) {
+                            if (simpleInventory.containsAny(ImmutableSet.of(Items.PUMPKIN_SEEDS))) {
+                                continue;
+                            }
+                        }
+                        // Plant melon seeds over melon slices if possible.
+                        if (CONFIG.villagersProfessionConfig.farmersHarvestPumpkins && itemStack.isOf(Items.MELON_SLICE)) {
+                            if (simpleInventory.containsAny(ImmutableSet.of(Items.MELON_SEEDS))) {
+                                continue;
+                            }
+                        }
+                        if (!itemStack.isEmpty()) {
+                            if (plantSeed(itemStack, i, serverWorld, villagerEntity)) {
+                                break;
+                            }
+                        }
                     }
                 }
             }
 
-            this.ticksRan++;
+            this.targetPositions.remove(currentTarget);
+            this.currentTarget = null;
         }
+
+        if (this.currentTarget == null) {
+            this.currentTarget = this.chooseRandomTarget(serverWorld);
+        }
+
+        if (this.currentTarget != null) {
+            this.nextResponseTime = l + 20L;
+            int completionRange = 1;
+            if (serverWorld.getBlockState(this.currentTarget).getBlock() instanceof GourdBlock) {
+                completionRange = 2;
+            }
+            villagerEntity.getBrain().remember(MemoryModuleType.WALK_TARGET,
+                    new WalkTarget(new BlockPosLookTarget(this.currentTarget), 0.5F, completionRange));
+            villagerEntity.getBrain()
+                    .remember(MemoryModuleType.LOOK_TARGET, new BlockPosLookTarget(this.currentTarget));
+        }
+
+        this.ticksRan++;
     }
 
     private boolean plantSeed(ItemStack itemStack, int stackIndex, ServerWorld serverWorld,
@@ -176,6 +222,16 @@ public abstract class FarmerVillagerTaskMixin extends Task<VillagerEntity> {
             blockState2 = Blocks.CARROTS.getDefaultState();
         } else if (itemStack.isOf(Items.BEETROOT_SEEDS)) {
             blockState2 = Blocks.BEETROOTS.getDefaultState();
+        } else if (itemStack.isOf(Items.MELON_SEEDS) || itemStack.isOf(Items.MELON_SLICE)) {
+            // Melon slices convert 1:1 to melon seeds.
+            blockState2 = Blocks.MELON_STEM.getDefaultState();
+        } else if (itemStack.isOf(Items.PUMPKIN_SEEDS) || itemStack.isOf(Items.PUMPKIN)) {
+            if (itemStack.isOf(Items.PUMPKIN)) {
+                // Give the remaining seeds to the farmer, the fourth is used for planting.
+                ItemStack pumpkinSeedItemStack = new ItemStack(Items.PUMPKIN_SEEDS, 3);
+                villagerEntity.getInventory().addStack(pumpkinSeedItemStack);
+            }
+            blockState2 = Blocks.PUMPKIN_STEM.getDefaultState();
         } else {
             return false;
         }
@@ -195,6 +251,12 @@ public abstract class FarmerVillagerTaskMixin extends Task<VillagerEntity> {
         BlockState blockState = world.getBlockState(pos);
         Block block = blockState.getBlock();
         Block block2 = world.getBlockState(pos.down()).getBlock();
+        if (CONFIG.villagersProfessionConfig.farmersHarvestMelons && block instanceof MelonBlock) {
+            return true;
+        }
+        if (CONFIG.villagersProfessionConfig.farmersHarvestPumpkins && block instanceof PumpkinBlock) {
+            return true;
+        }
         return ((block instanceof CropBlock && ((CropBlock) block).isMature(blockState)) ||
                 (blockState.isAir() && block2 instanceof FarmlandBlock && villagerEntity.hasSeedToPlant()));
     }
@@ -202,7 +264,10 @@ public abstract class FarmerVillagerTaskMixin extends Task<VillagerEntity> {
     @Inject(method = "shouldKeepRunning", at = @At(value = "HEAD"), cancellable = true)
     protected void replaceShouldKeepRunning(ServerWorld serverWorld, VillagerEntity villagerEntity, long l,
                                             CallbackInfoReturnable<Boolean> cir) {
-        cir.setReturnValue(this.currentTarget != null);
+        if (this.currentTarget == null) {
+            this.currentTarget = this.chooseRandomTarget(serverWorld);
+        }
+        cir.setReturnValue(this.currentTarget != null && this.ticksRan < MAX_RUN_TIME);
         cir.cancel();
     }
 }