+++ src/main/java/com/gitsh01/libertyvillagers/mixin/FindPointOfInterestTaskMixin.java
@@ -1,20 +1,30 @@
 package com.gitsh01.libertyvillagers.mixin;
 
 import com.google.common.collect.ImmutableMap;
+import net.minecraft.block.BedBlock;
+import net.minecraft.block.BlockState;
 import net.minecraft.entity.EntityType;
+import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.ai.brain.MemoryModuleType;
 import net.minecraft.entity.ai.brain.task.FindPointOfInterestTask;
 import net.minecraft.entity.ai.brain.task.Task;
 import net.minecraft.entity.mob.PathAwareEntity;
 import net.minecraft.server.world.ServerWorld;
+import net.minecraft.tag.BlockTags;
 import net.minecraft.util.dynamic.GlobalPos;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.world.poi.PointOfInterestStorage;
+import net.minecraft.world.poi.PointOfInterestType;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.ModifyArg;
+import org.spongepowered.asm.mixin.injection.Redirect;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
+
+import java.util.function.Predicate;
+import java.util.stream.Stream;
 
 import static com.gitsh01.libertyvillagers.LibertyVillagersMod.CONFIG;
 
@@ -22,6 +32,7 @@ import static com.gitsh01.libertyvillagers.LibertyVillagersMod.CONFIG;
 public abstract class FindPointOfInterestTaskMixin extends Task<PathAwareEntity> {
 
     private static final long TIME_NIGHT = 13000;
+    private ServerWorld world;
     private PathAwareEntity entity;
     @Shadow
     private MemoryModuleType<GlobalPos> targetMemoryModuleType;
@@ -49,18 +60,28 @@ public abstract class FindPointOfInterestTaskMixin extends Task<PathAwareEntity>
             at = @At("Head"))
     private void getLivingEntityForRun(ServerWorld world, PathAwareEntity entity, long l, CallbackInfo ci) {
         this.entity = entity;
+        this.world = world;
     }
 
-    @ModifyArg(method = "run(Lnet/minecraft/server/world/ServerWorld;Lnet/minecraft/entity/mob/PathAwareEntity;J)V",
+    @Redirect(method = "run(Lnet/minecraft/server/world/ServerWorld;Lnet/minecraft/entity/mob/PathAwareEntity;J)V",
             at = @At(value = "INVOKE",
-                    target = "Lnet/minecraft/world/poi/PointOfInterestStorage;" + "getSortedPositions" +
-                            "(Ljava/util/function/Predicate;Ljava/util/function/Predicate;Lnet/minecraft/util/math/BlockPos;" +
-                            "ILnet/minecraft/world/poi/PointOfInterestStorage$OccupationStatus;)" +
-                            "Ljava/util/stream/Stream;"), index = 3)
-    protected int modifyRunGetSortedPositionsArgs(int range) {
-        if (this.entity.getType() == EntityType.VILLAGER) {
-            return CONFIG.villagersGeneralConfig.findPOIRange;
-        }
-        return range;
+                    target = "Lnet/minecraft/world/poi/PointOfInterestStorage;getSortedPositions(Ljava/util/function/Predicate;Ljava/util/function/Predicate;Lnet/minecraft/util/math/BlockPos;ILnet/minecraft/world/poi/PointOfInterestStorage$OccupationStatus;)Ljava/util/stream/Stream;"))
+    public Stream<BlockPos> modifyGetSortedPositions(
+            PointOfInterestStorage pointOfInterestStorage, Predicate<PointOfInterestType> typePredicate,
+            Predicate<BlockPos> posPredicate, BlockPos pos, int radius,
+            PointOfInterestStorage.OccupationStatus occupationStatus) {
+        Predicate<BlockPos> newBlockPosPredicate = blockPos -> {
+            if (isBedOccupiedByOthers(this.world, blockPos, this.entity)) {
+                return false;
+            }
+            return posPredicate.test(blockPos);
+        };
+        return pointOfInterestStorage.getSortedPositions(typePredicate, newBlockPosPredicate, pos,
+                CONFIG.villagersGeneralConfig.findPOIRange, PointOfInterestStorage.OccupationStatus.HAS_SPACE);
+    }
+
+    private boolean isBedOccupiedByOthers(ServerWorld world, BlockPos pos, LivingEntity entity) {
+        BlockState blockState = world.getBlockState(pos);
+        return blockState.isIn(BlockTags.BEDS) && blockState.get(BedBlock.OCCUPIED) && !entity.isSleeping();
     }
 }