+++ src/main/java/com/gitsh01/libertyvillagers/mixin/WalkHomeTaskMixin.java
@@ -1,17 +1,23 @@
 package com.gitsh01.libertyvillagers.mixin;
 
 import com.google.common.collect.ImmutableMap;
+import net.minecraft.block.BedBlock;
+import net.minecraft.block.BlockState;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.ai.brain.task.Task;
 import net.minecraft.entity.ai.brain.task.WalkHomeTask;
+import net.minecraft.server.world.ServerWorld;
+import net.minecraft.tag.BlockTags;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.Vec3i;
 import net.minecraft.world.poi.PointOfInterestStorage;
 import net.minecraft.world.poi.PointOfInterestType;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.ModifyArgs;
 import org.spongepowered.asm.mixin.injection.Redirect;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.invoke.arg.Args;
 
 import java.util.function.Predicate;
@@ -21,6 +27,9 @@ import static com.gitsh01.libertyvillagers.LibertyVillagersMod.CONFIG;
 
 @Mixin(WalkHomeTask.class)
 public abstract class WalkHomeTaskMixin extends Task<LivingEntity> {
+
+    private ServerWorld world;
+    private LivingEntity entity;
 
     public WalkHomeTaskMixin() {
         super(ImmutableMap.of());
@@ -42,6 +51,13 @@ public abstract class WalkHomeTaskMixin extends Task<LivingEntity> {
         args.set(3, PointOfInterestStorage.OccupationStatus.HAS_SPACE);
     }
 
+    @Inject(method = "run(Lnet/minecraft/server/world/ServerWorld;Lnet/minecraft/entity/LivingEntity;J)V",
+        at = @At("HEAD"))
+    protected void runHead(ServerWorld world, LivingEntity entity, long time, CallbackInfo ci) {
+        this.world = world;
+        this.entity = entity;
+    }
+
     @Redirect(method = "run(Lnet/minecraft/server/world/ServerWorld;Lnet/minecraft/entity/LivingEntity;J)V",
             at = @At(value = "INVOKE", target = "Lnet/minecraft/world/poi/PointOfInterestStorage;getPositions" +
                     "(Ljava/util/function/Predicate;Ljava/util/function/Predicate;Lnet/minecraft/util/math/BlockPos;" +
@@ -50,7 +66,20 @@ public abstract class WalkHomeTaskMixin extends Task<LivingEntity> {
             PointOfInterestStorage pointOfInterestStorage, Predicate<PointOfInterestType> typePredicate,
             Predicate<BlockPos> posPredicate, BlockPos pos, int radius,
             PointOfInterestStorage.OccupationStatus occupationStatus) {
-        return pointOfInterestStorage.getSortedPositions(typePredicate, posPredicate, pos,
+        Predicate<BlockPos> newBlockPosPredicate = blockPos -> {
+            if (isBedOccupiedByOthers(this.world, blockPos, this.entity)) {
+                System.out.printf("%s Ignoring bed %s because it is occupied.\n",
+                        this.entity.getCustomName(), blockPos.toShortString());
+                return false;
+            }
+            return posPredicate.test(blockPos);
+        };
+        return pointOfInterestStorage.getSortedPositions(typePredicate, newBlockPosPredicate, pos,
                 CONFIG.villagersGeneralConfig.findPOIRange, PointOfInterestStorage.OccupationStatus.HAS_SPACE);
+    }
+
+    private boolean isBedOccupiedByOthers(ServerWorld world, BlockPos pos, LivingEntity entity) {
+        BlockState blockState = world.getBlockState(pos);
+        return blockState.isIn(BlockTags.BEDS) && blockState.get(BedBlock.OCCUPIED) && !entity.isSleeping();
     }
 }